WATrailsRecommender/Microservice/requirements.txt
numpy==1.20.1
pandas==1.2.4
Pillow==8.4.0
streamlit==1.2.0
